Moving Unit tests' Assert3 to its own file.
